% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze}
\alias{analyze}
\title{analyze}
\usage{
analyze(data, y, x, title, success = "Traditional", scope, conf = 0.95,
  alpha = 0.05)
}
\arguments{
\item{data}{Data frame containing two columns; the response and the explanatory variable.}

\item{y}{Character string label for the response variable.}

\item{x}{Character string label for the explanatory variable.}

\item{title}{Character string containing title of the plot}

\item{success}{Character string equivalent to the response variable level to be analyzed.}

\item{scope}{Character string indicating the scope of the data.}

\item{conf}{Numeric between 0 and 1, indicates desired confidence level for hypothesis test.}

\item{alpha}{Numeric between 0 and 1, indicates the target probability of a type I error.}
}
\value{
list containing three types of objects: (1) data, both raw and summarized observed and expected frequencies,
(2) plots, observed and expected frequencies and proportions in bar plot format, and (3), tests the results of the
independence and proportion tests.
group a
}
\description{
\code{results} Performs an analysis for a data frame containing a response
and an explanatory variable.  The key functions are:
\describe{
 \item{formatData(d, y, x)}{Takes three variables: d, a data frame containing response and explanatory variable;
 y, a label for the response variable; and x, a label for the explanatory variable, and returns the original raw data
 as well as the summarized observed and expected frequencies.}
 \item{x2(d, y, x, conf, alpha)}{Takes four variables: d, a data frame containing response and explanatory variable;,
 y, a label for the response variable; x, a label for the explanatory variable; conf numeric between 0 and 1 that indicates
 the confidence level; and alpha a number between 0 and 1 that indicates the probability of a type I error. It returns the
 results of a chi-square test of independence with plot and descriptive statements regarding the result.}
 \item{zTest(d, success, order, alternative, conf, alpha)}{Takes six variables: d, a data frame containing response and explanatory variable;
 success is the level of the response variable considered for analysis; order is the order of the levels for the explanatory variable;,
 alternative is the alternative hypothesis for the test; conf numeric between 0 and 1 that indicates
 the confidence level; and alpha a number between 0 and 1 that indicates the probability of a type I error. It returns the results
 of a two-proportion z-test with descriptive statements.}
 \item{propTest(d, order, alternative, success, conf, alpha)}{Takes six variables: d, a data frame containing response and explanatory variable;
 success is the level of the response variable considered for analysis; order is the order of the levels for the explanatory variable;,
 alternative is the alternative hypothesis for the test; conf numeric between 0 and 1 that indicates
 the confidence level; and alpha a number between 0 and 1 that indicates the probability of a type I error. It returns the results
 of a two-proportion z-test with descriptive statements.}
}
}
\seealso{
Other xmar functions: \code{\link{X23D}}, \code{\link{X2}},
  \code{\link{analyze3d}}, \code{\link{discussion}},
  \code{\link{formatData}}, \code{\link{plotBars}},
  \code{\link{preprocess}}, \code{\link{propTest}},
  \code{\link{univariate}}, \code{\link{zTest}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
