% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze3d.R
\name{analyze3d}
\alias{analyze3d}
\title{analyze3d}
\usage{
analyze3d(data2d, data3d)
}
\arguments{
\item{data}{Data frame containing two columns; the response and the explanatory variable.}
}
\description{
\code{results} Performs an analysis for a data frame containing a response
and an explanatory variable.  The key functions are:
}
\seealso{
Other xmar functions: \code{\link{X23D}}, \code{\link{X2}},
  \code{\link{analyze}}, \code{\link{formatData}},
  \code{\link{plotBars}}, \code{\link{preprocess}},
  \code{\link{propTest}}, \code{\link{univariate}},
  \code{\link{zTest}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
