% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discussion.R
\name{discussion}
\alias{discussion}
\title{discussion}
\usage{
discussion(analysis)
}
\arguments{
\item{analysis}{Data frame containing the results of analysis and hypothesis tests.}
}
\value{
list containing a data frame of summarized data and a plot.
}
\description{
\code{discussion} Extracts data and renders plots for discussion section.
}
\seealso{
Other xmar functions: \code{\link{X23D}}, \code{\link{X2}},
  \code{\link{analyze3d}}, \code{\link{analyze}},
  \code{\link{formatData}}, \code{\link{plotBars}},
  \code{\link{preprocess}}, \code{\link{propTest}},
  \code{\link{univariate}}, \code{\link{zTest}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
