% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBars.R
\name{plotBars}
\alias{plotBars}
\title{plotBars}
\usage{
plotBars(data, y, x)
}
\arguments{
\item{data}{Data frame containing proportions comprised of two columns: 'target' and 'group'.
'target' represents the proportions for the target variable and 'group' is the name of the grouping variable}

\item{y}{Character string indicating the name of the response variable.}

\item{x}{Character string indicating the name of the explanatory variable.}
}
\description{
\code{plotBars} Renders segmented bar plots with value labels representing
proportions each segment represents.
}
\seealso{
Other xmar functions: \code{\link{X2}},
  \code{\link{analyze}}, \code{\link{formatData}},
  \code{\link{preprocess}}, \code{\link{propTest}},
  \code{\link{univariate}}, \code{\link{zTest}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
