% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rq1.R
\name{rq1}
\alias{rq1}
\title{rq1}
\usage{
rq1(xmar)
}
\description{
\code{rq1} Performs analysis for research question 1: Are attitudes about
and behaviors in extra-marital sex associated.
}
\seealso{
Other xmar functions: \code{\link{preprocess}},
  \code{\link{sampleSize}}, \code{\link{univariate}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
