% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rq1a.R
\name{rq1a}
\alias{rq1a}
\title{rq1a}
\usage{
rq1a(xmar)
}
\description{
\code{rq1a} Performs analysis for research question 1a: Are opinions about
and behaviors in extra-marital sex associated.
}
\seealso{
Other xmar functions: \code{\link{X2a}}, \code{\link{X2}},
  \code{\link{formatData}}, \code{\link{plotBars}},
  \code{\link{preprocess}}, \code{\link{rq1b}},
  \code{\link{rq2}}, \code{\link{rq3}},
  \code{\link{sampleSize}}, \code{\link{univariate}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
