% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rq1b.R
\name{rq1b}
\alias{rq1b}
\title{rq1b}
\usage{
rq1b(rq1aData)
}
\description{
\code{rq1b} Performs analysis for research question 1b: Are the proportions
of the population that have engaged in extra-marital conduct the same for
those who believe such conduct is "sometimes wrong" and those who
consider such conduct to be "not wront at all".
}
\seealso{
Other xmar functions: \code{\link{X2a}}, \code{\link{X2}},
  \code{\link{formatData}}, \code{\link{plotBars}},
  \code{\link{preprocess}}, \code{\link{rq1a}},
  \code{\link{rq2}}, \code{\link{rq3}},
  \code{\link{sampleSize}}, \code{\link{univariate}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
