% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rq2.R
\name{rq2}
\alias{rq2}
\title{rq2}
\usage{
rq2(xmar)
}
\description{
\code{rq2} Performs analysis for research question 2: Are opinions associated
with region, political view, class, or education.
}
\seealso{
Other xmar functions: \code{\link{X2}},
  \code{\link{formatData}}, \code{\link{plotBars}},
  \code{\link{preprocess}}, \code{\link{rq1}},
  \code{\link{rq3}}, \code{\link{sampleSize}},
  \code{\link{univariate}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
