% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rq3.R
\name{rq3}
\alias{rq3}
\title{rq3}
\usage{
rq3(xmar)
}
\description{
\code{rq3} Performs analysis for research question 2: Are opinions associated
with region, political view, class, or education.
}
\seealso{
Other xmar functions: \code{\link{X2a}}, \code{\link{X2}},
  \code{\link{formatData}}, \code{\link{plotBars}},
  \code{\link{preprocess}}, \code{\link{rq1a}},
  \code{\link{rq1b}}, \code{\link{rq2}},
  \code{\link{sampleSize}}, \code{\link{univariate}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
