% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSize.R
\name{sampleSize}
\alias{sampleSize}
\title{sampleSize}
\usage{
sampleSize(xmar)
}
\description{
\code{sampleSize} Performs a sampleSize analysis for several null proportions based
upon the number of levels in each grouping variable.
}
\seealso{
Other xmar functions: \code{\link{preprocess}},
  \code{\link{rq1}}, \code{\link{univariate}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
