% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R, R/univariate.R
\name{univariate}
\alias{univariate}
\alias{univariate}
\title{univariate}
\usage{
univariate(xmar)

univariate(xmar)
}
\description{
\code{univariate} Performs univariate analysis of variables

\code{univariate} Performs univariate analysis of variables
}
\seealso{
Other xmar functions: \code{\link{X2a}}, \code{\link{X2}},
  \code{\link{formatData}}, \code{\link{plotBars}},
  \code{\link{preprocess}}, \code{\link{rqa}}

Other xmar functions: \code{\link{X2a}}, \code{\link{X2}},
  \code{\link{formatData}}, \code{\link{plotBars}},
  \code{\link{preprocess}}, \code{\link{rqa}}
}
\author{
John James, \email{jjames@datasciencesalon.org}

John James, \email{jjames@datasciencesalon.org}
}
