% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X2a.R
\name{X2a}
\alias{X2a}
\title{X2a}
\usage{
X2a(x2aTable, target, group, p = 0.05)
}
\arguments{
\item{x2aTable}{Contingency table object containing observed values for chi-square hypothesis test}

\item{target}{Character string indicating the name of target or study variable}

\item{group}{Character string indicating the name of grouping variable}

\item{p}{Alpha .the probability of a type 1 error}
}
\description{
\code{X2a} Conducts a chi-square test of independency on a contingency table
and renders a data frame containing the results and a plot showing the
chi-square distribution with reject region and observed test statistic.
}
\seealso{
Other xmar functions: \code{\link{X2}},
  \code{\link{formatData}}, \code{\link{plotBars}},
  \code{\link{preprocess}}, \code{\link{rqa}},
  \code{\link{univariate}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
