% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zTest.R
\name{zTest}
\alias{zTest}
\title{zTest}
\usage{
zTest(data, y, x, order = NULL, alternative = "two.sided",
  success = "Liberal", conf = 0.95)
}
\arguments{
\item{data}{data frame containing data to be analyzed}

\item{y}{Chracter vector containing the name of the response variable}

\item{x}{Chracter vector containing the name of the explanatory variable}

\item{order}{when x is given, order of levels of x in which to subtract parameters}

\item{alternative}{direction of the alternative hypothesis; "less","greater", or "two.sided"}

\item{conf}{confidence level, value between 0 and 1}
}
\value{
Data frame containing the results of the z-test
}
\description{
\code{zTest} Conducts a z-tests for a multi-level categorical explanatory
variable on a response variable. Returns data and plots of the z-distribution
vis-a-vis the observed test statistic.
}
\seealso{
Other xmar functions: \code{\link{X2}},
  \code{\link{formatData}}, \code{\link{plotBars}},
  \code{\link{preprocess}}, \code{\link{rqa}},
  \code{\link{univariate}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
