% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatData.R
\name{formatData}
\alias{formatData}
\title{formatData}
\usage{
formatData(data)
}
\arguments{
\item{data}{Data frame containing two or three columns, the values for the
response and explanatory variable(s). The first column is the response
variable.}
}
\value{
List containing a contingency table and a data frame of descriptive statistics
}
\description{
\code{formatData} Render frequency and proportion data
}
\seealso{
Other xmar functions: \code{\link{X23D}}, \code{\link{X2}},
  \code{\link{analyze3d}}, \code{\link{analyze}},
  \code{\link{discussion}}, \code{\link{plotBars}},
  \code{\link{preprocess}}, \code{\link{propTest}},
  \code{\link{univariate}}, \code{\link{zTest}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
