% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatData.R
\name{formatData}
\alias{formatData}
\title{formatData}
\usage{
formatData(data, y, x)
}
\arguments{
\item{data}{Data frame containing two columns, the values for the response and explanatory variables}

\item{y}{Character string indicating the description of the response variable}

\item{x}{Character string indicating the description of the explanatory variable}
}
\value{
List containing a contingency table and a data frame of descriptive statistics
}
\description{
\code{formatData} Render frequency and proportion data
}
\seealso{
Other xmar functions: \code{\link{X2}},
  \code{\link{analyze}}, \code{\link{plotBars}},
  \code{\link{preprocess}}, \code{\link{propTest}},
  \code{\link{univariate}}, \code{\link{zTest}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
