% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatData.R
\name{formatData}
\alias{formatData}
\title{formatData}
\usage{
formatData(data, target, group)
}
\arguments{
\item{data}{Data frame containing proportions comprised of two columns: 'target' and 'group'.
'target' represents the proportions for the target variable and 'group' is the name of the grouping variable}

\item{target}{Character string indicating the name of the target variable}

\item{group}{Character string indicating the name of the grouping variable}
}
\description{
\code{formatData} Render frequency and proportion data
}
\seealso{
Other xmar functions: \code{\link{X2a}}, \code{\link{X2}},
  \code{\link{plotBars}}, \code{\link{preprocess}},
  \code{\link{rq1a}}, \code{\link{rq1b}},
  \code{\link{rq2}}, \code{\link{rq3}},
  \code{\link{sampleSize}}, \code{\link{univariate}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
