% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBars.R
\name{plotBars}
\alias{plotBars}
\title{plotBars}
\usage{
plotBars(data, target, group)
}
\arguments{
\item{data}{Data frame containing proportions comprised of two columns: 'target' and 'group'.
'target' represents the proportions for the target variable and 'group' is the name of the grouping variable}

\item{target}{Character string indicating the name of the target variable.}

\item{group}{Character string indicating the name of the grouping variable.}
}
\description{
\code{plotBars} Renders segmented bar plots with value labels representing
proportions each segment represents.
}
\seealso{
Other xmar functions: \code{\link{X2}},
  \code{\link{formatData}}, \code{\link{preprocess}},
  \code{\link{rq1}}, \code{\link{rq2}}, \code{\link{rq3}},
  \code{\link{sampleSize}}, \code{\link{univariate}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
