% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{preprocess}
\usage{
preprocess(GSS)
}
\description{
\code{preprocess} Creates additional variables required for analysis.
}
\seealso{
Other xmar functions: \code{\link{X2}},
  \code{\link{formatData}}, \code{\link{plotBars}},
  \code{\link{rqa}}, \code{\link{univariate}},
  \code{\link{zTest}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
