% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqa.R
\name{rqa}
\alias{rqa}
\title{rqa}
\usage{
rqa(xmar)
}
\arguments{
\item{xmar}{Data frame containing the study data}
}
\value{
list containing the results of hypothesis tests forresearch question
group a
}
\description{
\code{results} Performs the analysis for research question group a, which
includes reearch questions 1-3.
}
\seealso{
Other xmar functions: \code{\link{X2}},
  \code{\link{formatData}}, \code{\link{plotBars}},
  \code{\link{preprocess}}, \code{\link{univariate}},
  \code{\link{zTest}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
