% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{univariate}
\alias{univariate}
\title{univariate}
\usage{
univariate(xmar)
}
\description{
\code{univariate} Performs univariate analysis of variables
}
\seealso{
Other xmar functions: \code{\link{X2a}}, \code{\link{X2}},
  \code{\link{formatData}}, \code{\link{plotBars}},
  \code{\link{preprocess}}, \code{\link{rq1a}},
  \code{\link{rq1b}}, \code{\link{rq2}}, \code{\link{rq3}},
  \code{\link{sampleSize}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
