% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X2.R
\name{X2}
\alias{X2}
\title{X2}
\usage{
X2(x2Table, y, x, p = 0.05)
}
\arguments{
\item{x2Table}{Contingency table object containing observed values for chi-square hypothesis test}

\item{y}{Character string indicating the name of the respnse variable}

\item{x}{Character string indicating the name of the explanatory variable}

\item{p}{Alpha .the probability of a type 1 error}
}
\description{
\code{X2} Conducts a chi-square test of independency on a contingency table
and renders a data frame containing the results and a plot showing the
chi-square distribution with reject region and observed test statistic.
}
\seealso{
Other xmar functions: \code{\link{formatData}},
  \code{\link{plotBars}}, \code{\link{preprocess}},
  \code{\link{rqa}}, \code{\link{univariate}},
  \code{\link{zTest}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
