% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X2.R
\name{X2}
\alias{X2}
\title{X2}
\usage{
X2(table, target, group, p = 0.05)
}
\arguments{
\item{table}{Contingency table object containing observed values for chi-square hypothesis test}

\item{target}{Character string indicating the name of target or study variable}

\item{group}{Character string indicating the name of grouping variable}

\item{p}{Alpha .the probability of a type 1 error}
}
\description{
\code{X2} Conducts a chi-square test of independency on a contingency table
and renders a data frame containing the results and a plot showing the
chi-square distribution with reject region and observed test statistic.
}
\seealso{
Other xmar functions: \code{\link{formatData}},
  \code{\link{plotBars}}, \code{\link{preprocess}},
  \code{\link{rq1}}, \code{\link{rq2}}, \code{\link{rq3}},
  \code{\link{sampleSize}}, \code{\link{univariate}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
